+++ src/main/java/com/ticle/server/user/dto/EmailDto.java
@@ -1,0 +1,9 @@
+package com.ticle.server.user.dto;
+
+import lombok.Data;
+
+@Data
+public class EmailDto {
+    private String mail;
+    private String code;
+}
